requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
mpmath==1.3.0
mysql-connector-python==9.4.0
requests==2.32.5
sympy==1.14.0
urllib3==2.5.0
